Add player entity that you can move around the canvas and which animates.

+++ src/scenes/invasion/components/position.ts
@@ -1,8 +1,8 @@
 import { MAX_ENTITIES } from "@src/framework/constants";
 
 const Position = {
-  x: new Int16Array(MAX_ENTITIES),
-  y: new Int16Array(MAX_ENTITIES),
+  x: new Float32Array(MAX_ENTITIES),
+  y: new Float32Array(MAX_ENTITIES),
   w: new Int16Array(MAX_ENTITIES),
   h: new Int16Array(MAX_ENTITIES),
 };
